Processing updates
Reduced the input to one button, not two. Formatted things better in the pic div. Started working on making a displayable dog breed name by processing the URL of the picture (containing the breed name) with multiple regexes. Messy at the moment, but can be streamlined later once it's functional